Directory refactoring
