+++ include/NodeSystem/Node.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Global.hpp"
+#include "../Global.hpp"
 
 class NodeInput;
 class NodeOutput;
