--- src/Node.cpp
@@ -1,6 +1,0 @@
-#include "include/Node.hpp"
-
-Node::Node()
-{
-
-}
