--- src/NodeInput.cpp
@@ -1,6 +1,0 @@
-#include "include/NodeInput.hpp"
-
-NodeInput::NodeInput()
-{
-
-}
