--- src/NodeOutput.cpp
@@ -1,6 +1,0 @@
-#include "include/NodeOutput.hpp"
-
-NodeOutput::NodeOutput()
-{
-
-}
