+++ src/NodeSystem/Node.cpp
@@ -1,0 +1,6 @@
+#include "include/NodeSystem/Node.hpp"
+
+Node::Node()
+{
+
+}
