+++ src/NodeSystem/NodeInput.cpp
@@ -1,0 +1,6 @@
+#include "include/NodeSystem/NodeInput.hpp"
+
+NodeInput::NodeInput()
+{
+
+}
