+++ src/NodeSystem/NodeOutput.cpp
@@ -1,0 +1,6 @@
+#include "include/NodeSystem/NodeOutput.hpp"
+
+NodeOutput::NodeOutput()
+{
+
+}
